Adds fightkeyb keyboard
This is a handwired, hitbox-like board to be used in fighting games. The
default layout mirrors the default Street Fighter V keyboard scheme.

There's also a layout using arrow keys instead of ASDW. The layers are
persistent so if you change the layout it will stay there even if
disconnected.

This board is working from the a-chol fork for joystick support, and I
will add some axis and joystick keycodes to a third layout in the future
(the joystick support still experimental and failing on windows).